freesound: try to fix built-in oauth hack
Fix build with OAUTH_BUILTIN_HACK. Sadly, this doesn't work with the
current freesound.org web site. For reasons I can't fathom, the page
returned by the call to curl_easy_perform() at line 520 is not the same
as the page returned from the exact same URL to a browser: it contains
neither the authorization token, nor the "Authorize!" button.